Further cleanup and new examples util file
